+++ src/src/Model/SubscriberCollection.php
@@ -25,6 +25,9 @@ declare(strict_types=1);
 
 namespace PrestaShop\Module\SmailyForPrestaShop\Model;
 
+use Db;
+use DbQuery;
+
 if (!defined('_PS_VERSION_')) {
     exit;
 }
@@ -44,7 +47,22 @@ class SubscriberCollection
             )
         ) . ')';
 
-        return \Db::getInstance()->execute($query);
+        if ( Db::getInstance()->execute($query) !== true ) {
+            return false;
+        };
+
+        $query = 'UPDATE ' . _DB_PREFIX_ . 'emailsubscription SET active=0 WHERE email IN (' .
+        implode(
+            ', ',
+            array_map(
+                function ($email) {
+                    return "'" . pSQL($email) . "'";
+                },
+                $emails
+            )
+        ) . ')';
+
+        return Db::getInstance()->execute($query);
     }
 
     /**
@@ -54,22 +72,41 @@ class SubscriberCollection
      */
     public function getSubscribers(int $limit, int $offset = 0): array
     {
-        $sql = new \DbQuery();
-        $sql->select('*');
+        $subscribers = [];
+
+        $sql = new DbQuery();
+        $sql->select('s.`domain` AS `shop_domain`, c.`lastname`, c.`firstname`, c.`email`, c.`birthday`, c.`website`');
         $sql->from('customer', 'c');
         $sql->where('c.newsletter = 1');
+        $sql->leftJoin('shop_url', 's', 's.id_shop = c.id_shop');
         $sql->limit($limit, $offset);
 
-        $result = \Db::getInstance()->executeS($sql);
+        $customers = Db::getInstance()->executeS($sql);
 
-        $subscribers = [];
-        foreach ($result as $subscriber) {
+        foreach ($customers as $subscriber) {
             $s = new Subscriber();
             $s->email = $subscriber['email'];
             $s->firstName = $subscriber['firstname'];
             $s->lastName = $subscriber['lastname'];
             $s->birthDay = $subscriber['birthday'];
             $s->website = $subscriber['website'];
+            $s->store_url = $subscriber['shop_domain'];
+
+            $subscribers[] = $s;
+        }
+
+        $sql = new DbQuery();
+        $sql->select('e.`email`, s.`domain` AS `shop_domain`');
+        $sql->from('emailsubscription', 'e');
+        $sql->where('e.active = 1');
+        $sql->leftJoin('shop_url', 's', 's.id_shop = e.id_shop');
+        $sql->limit($limit, $offset);
+
+        $non_customers = Db::getInstance()->executeS($sql);
+        foreach ($non_customers as $subscriber) {
+            $s = new Subscriber();
+            $s->email = $subscriber['email'];
+            $s->store_url = $subscriber['shop_domain'];
 
             $subscribers[] = $s;
         }